requirements.txt
google-api-python-client
vaderSentiment
matplotlib
python-dotenv
emoji
